// frontend/src/app/models/evaluation.ts
export interface Evaluation {
  id?: number;
  clariteDesObjectifs: string;
  pertinenceDesSupportsPedagogiques: string;
  competencesDuFormateur: string;
  reponseAuxAttentes: string;
  organisationGenerale: string;
  participant?: any; // Replace with the correct type if needed
  cycle?: any; // Replace with the correct type if needed
}